enable to attach/detach sources via my channels
