+++ linen-modern/src/main/scala/x7c1/linen/modern/init/settings/my/MyChannelSourcesDelegatee.scala
@@ -1,24 +1,30 @@
 package x7c1.linen.modern.init.settings.my
 
-import android.app.Activity
+import android.support.v4.app.FragmentActivity
+import android.support.v7.app.AlertDialog
 import android.support.v7.widget.LinearLayoutManager
 import x7c1.linen.glue.activity.ActivityControl
-import x7c1.linen.glue.res.layout.{SettingChannelSourcesLayout, SettingChannelSourcesRow}
+import x7c1.linen.glue.res.layout.{SettingChannelSourcesLayout, SettingChannelSourcesRow, SettingSourceAttach, SettingSourceAttachRowItem}
 import x7c1.linen.glue.service.ServiceControl
 import x7c1.linen.glue.service.ServiceLabel.Updater
 import x7c1.linen.modern.accessor.preset.ClientAccount
 import x7c1.linen.modern.accessor.{LinenOpenHelper, SettingSourceAccessorFactory}
-import x7c1.linen.modern.display.settings.{SourceMenuSelected, ChannelSourcesSelected, OnSyncClickedListener, SourceRowAdapter}
+import x7c1.linen.modern.display.settings.{ChannelSourcesSelected, OnSyncClickedListener, SourceRowAdapter}
+import x7c1.linen.modern.init.settings.source.OnSourceMenuSelected
 import x7c1.linen.modern.init.updater.UpdaterMethods
-import x7c1.wheat.ancient.resource.ViewHolderProvider
+import x7c1.wheat.ancient.context.ContextualFactory
+import x7c1.wheat.ancient.resource.{ViewHolderProvider, ViewHolderProviderFactory}
 import x7c1.wheat.macros.intent.{IntentExpander, ServiceCaller}
 import x7c1.wheat.macros.logger.Log
 import x7c1.wheat.modern.decorator.Imports._
 import x7c1.wheat.modern.resource.MetricsConverter
 
 class MyChannelSourcesDelegatee (
-  activity: Activity with ActivityControl with ServiceControl,
+  activity: FragmentActivity with ActivityControl with ServiceControl,
   layout: SettingChannelSourcesLayout,
+  dialogFactory: ContextualFactory[AlertDialog.Builder],
+  attachLayoutFactory: ViewHolderProviderFactory[SettingSourceAttach],
+  attachRowFactory: ViewHolderProviderFactory[SettingSourceAttachRowItem],
   sourceRowProvider: ViewHolderProvider[SettingChannelSourcesRow] ){
 
   private lazy val database =
@@ -46,7 +52,15 @@ class MyChannelSourcesDelegatee (
       account = ClientAccount(event.accountId),
       channelId = event.channelId,
       viewHolderProvider = sourceRowProvider,
-      onMenuSelected = new OnSourceMenuSelected().showMenu,
+      onMenuSelected = {
+        val listener = new OnSourceMenuSelected(
+          activity = activity,
+          dialogFactory = dialogFactory,
+          attachLayoutFactory = attachLayoutFactory,
+          attachRowFactory = attachRowFactory
+        )
+        listener.showMenu
+      },
       onSyncClicked = onSyncClicked,
       metricsConverter = MetricsConverter(activity)
     )
@@ -57,11 +71,5 @@ class MyChannelSourcesDelegatee (
       startService(activity, activity getClassOf Updater){
         _ loadSource event.sourceId
       }
-  }
-}
-
-class OnSourceMenuSelected {
-  def showMenu(event: SourceMenuSelected): Unit = {
-    Log info s"[init]"
   }
 }
