+++ linen-modern/src/main/scala/x7c1/linen/modern/init/settings/preset/PresetChannelSourcesDelegatee.scala
@@ -2,20 +2,18 @@ package x7c1.linen.modern.init.settings.preset
 
 import android.support.v4.app.FragmentActivity
 import android.support.v7.app.AlertDialog
-import android.support.v7.widget.PopupMenu.OnMenuItemClickListener
-import android.support.v7.widget.{PopupMenu, LinearLayoutManager}
-import android.view.{MenuItem, Menu}
+import android.support.v7.widget.LinearLayoutManager
 import x7c1.linen.glue.activity.ActivityControl
-import x7c1.linen.glue.res.layout.{SettingSourceAttachRowItem, SettingSourceAttach, SettingChannelSourcesLayout, SettingChannelSourcesRow}
+import x7c1.linen.glue.res.layout.{SettingChannelSourcesLayout, SettingChannelSourcesRow, SettingSourceAttach, SettingSourceAttachRowItem}
 import x7c1.linen.glue.service.ServiceControl
 import x7c1.linen.glue.service.ServiceLabel.Updater
 import x7c1.linen.modern.accessor.preset.ClientAccount
 import x7c1.linen.modern.accessor.{LinenOpenHelper, SettingSourceAccessorFactory}
-import x7c1.linen.modern.display.settings.{ChannelSourcesSelected, OnSyncClickedListener, SourceMenuSelected, SourceRowAdapter}
+import x7c1.linen.modern.display.settings.{ChannelSourcesSelected, OnSyncClickedListener, SourceRowAdapter}
+import x7c1.linen.modern.init.settings.source.OnSourceMenuSelected
 import x7c1.linen.modern.init.updater.UpdaterMethods
 import x7c1.wheat.ancient.context.ContextualFactory
-import x7c1.wheat.ancient.resource.{ViewHolderProviderFactory, ViewHolderProvider}
-import x7c1.wheat.macros.fragment.FragmentFactory
+import x7c1.wheat.ancient.resource.{ViewHolderProvider, ViewHolderProviderFactory}
 import x7c1.wheat.macros.intent.{IntentExpander, ServiceCaller}
 import x7c1.wheat.macros.logger.Log
 import x7c1.wheat.modern.decorator.Imports._
@@ -46,7 +44,6 @@ class PresetChannelSourcesDelegatee (
     database.close()
   }
   def showSources(event: ChannelSourcesSelected): Unit = {
-    Log info s"$event"
     val accessorFactory = new SettingSourceAccessorFactory(database, event.accountId)
     layout.sourceList setAdapter new SourceRowAdapter(
       accessor = accessorFactory create event.channelId,
@@ -69,38 +66,4 @@ class PresetChannelSourcesDelegatee (
       }
   }
 
-}
-
-class OnSourceMenuSelected(
-  activity: FragmentActivity,
-  dialogFactory: ContextualFactory[AlertDialog.Builder],
-  attachLayoutFactory: ViewHolderProviderFactory[SettingSourceAttach],
-  attachRowFactory: ViewHolderProviderFactory[SettingSourceAttachRowItem]
-){
-
-  def showMenu(event: SourceMenuSelected): Unit = {
-    Log info s"[init]"
-    val menu = new PopupMenu(activity, event.targetView)
-    menu.getMenu.add(Menu.NONE, 123, 1, "Attached channels")
-    menu setOnMenuItemClickListener new OnMenuItemClickListener {
-      override def onMenuItemClick(item: MenuItem): Boolean = {
-        if (item.getItemId == 123){
-          createAttachSourceDialog(event) showIn activity
-        }
-        true
-      }
-    }
-    menu.show()
-  }
-  def createAttachSourceDialog(event: SourceMenuSelected) = {
-    FragmentFactory.create[AttachSourceDialog] by
-      new AttachSourceDialog.Arguments(
-        clientAccountId = event.clientAccountId,
-        originalChannelId = event.channelId,
-        originalSourceId = event.selectedSourceId,
-        dialogFactory = dialogFactory,
-        attachLayoutFactory = attachLayoutFactory,
-        rowFactory = attachRowFactory
-      )
-  }
 }
